[manual impl] Remove "previous chunk available" optimization
As it turns out, calling wait_for(0s) on a future is not cost-free,
and by my measurements usually takes 45-50us, a very long time. The
amount saved by the optimization - in the fraction of times when it
can be used - is half that, at most.

This change buys another 7% or so in performance.